+++ ast/ast.go
@@ -61,6 +61,8 @@ const (
 	TagType
 	// CommentType type identifier for comment node
 	CommentType
+	// CommentGroupType type identifier for comment group node
+	CommentGroupType
 )
 
 // String node type identifier to text
@@ -106,6 +108,8 @@ func (t NodeType) String() string {
 		return "Tag"
 	case CommentType:
 		return "Comment"
+	case CommentGroupType:
+		return "CommentGroup"
 	}
 	return ""
 }
@@ -154,6 +158,8 @@ func (t NodeType) YAMLName() string {
 		return "tag"
 	case CommentType:
 		return "comment"
+	case CommentGroupType:
+		return "comment"
 	}
 	return ""
 }
@@ -170,9 +176,9 @@ type Node interface {
 	// AddColumn add column number to child nodes recursively
 	AddColumn(int)
 	// SetComment set comment token to node
-	SetComment(*token.Token) error
+	SetComment(*CommentGroupNode) error
 	// Comment returns comment token instance
-	GetComment() *token.Token
+	GetComment() *CommentGroupNode
 	// MarshalYAML
 	MarshalYAML() ([]byte, error)
 	// already read length
@@ -181,6 +187,8 @@ type Node interface {
 	addReadLen(int)
 	// clean read length
 	clearLen()
+	// String node to text without comment
+	stringWithoutComment() string
 }
 
 // ScalarNode type for scalar node
@@ -190,8 +198,12 @@ type ScalarNode interface {
 }
 
 type BaseNode struct {
-	Comment *token.Token
+	Comment *CommentGroupNode
 	read    int
+}
+
+func addCommentString(base string, node *CommentGroupNode) string {
+	return fmt.Sprintf("%s %s", base, node.String())
 }
 
 func (n *BaseNode) readLen() int {
@@ -207,16 +219,13 @@ func (n *BaseNode) addReadLen(len int) {
 }
 
 // GetComment returns comment token instance
-func (n *BaseNode) GetComment() *token.Token {
+func (n *BaseNode) GetComment() *CommentGroupNode {
 	return n.Comment
 }
 
 // SetComment set comment token
-func (n *BaseNode) SetComment(tk *token.Token) error {
-	if tk.Type != token.CommentType {
-		return ErrInvalidTokenType
-	}
-	n.Comment = tk
+func (n *BaseNode) SetComment(node *CommentGroupNode) error {
+	n.Comment = node
 	return nil
 }
 
@@ -400,7 +409,19 @@ func String(tk *token.Token) *StringNode {
 // Comment create node for comment
 func Comment(tk *token.Token) *CommentNode {
 	return &CommentNode{
-		BaseNode: &BaseNode{Comment: tk},
+		BaseNode: &BaseNode{},
+		Token:    tk,
+	}
+}
+
+func CommentGroup(comments []*token.Token) *CommentGroupNode {
+	nodes := []*CommentNode{}
+	for _, comment := range comments {
+		nodes = append(nodes, Comment(comment))
+	}
+	return &CommentGroupNode{
+		BaseNode: &BaseNode{},
+		Comments: nodes,
 	}
 }
 
@@ -522,6 +543,10 @@ func (f *File) String() string {
 	return strings.Join(docs, "\n")
 }
 
+func (f *File) stringWithoutComment() string {
+	return f.String()
+}
+
 // DocumentNode type of Document
 type DocumentNode struct {
 	*BaseNode
@@ -563,6 +588,10 @@ func (d *DocumentNode) String() string {
 	return strings.Join(doc, "\n")
 }
 
+func (d *DocumentNode) stringWithoutComment() string {
+	return d.String()
+}
+
 // MarshalYAML encodes to a YAML text
 func (d *DocumentNode) MarshalYAML() ([]byte, error) {
 	return []byte(d.String()), nil
@@ -575,8 +604,7 @@ func removeUnderScoreFromNumber(num string) string {
 // NullNode type of null node
 type NullNode struct {
 	*BaseNode
-	Comment *token.Token // position of Comment ( `#comment` )
-	Token   *token.Token
+	Token *token.Token
 }
 
 // Read implements (io.Reader).Read
@@ -597,15 +625,6 @@ func (n *NullNode) AddColumn(col int) {
 	n.Token.AddColumn(col)
 }
 
-// SetComment set comment token
-func (n *NullNode) SetComment(tk *token.Token) error {
-	if tk.Type != token.CommentType {
-		return ErrInvalidTokenType
-	}
-	n.Comment = tk
-	return nil
-}
-
 // GetValue returns nil value
 func (n *NullNode) GetValue() interface{} {
 	return nil
@@ -613,6 +632,13 @@ func (n *NullNode) GetValue() interface{} {
 
 // String returns `null` text
 func (n *NullNode) String() string {
+	if n.Comment != nil {
+		return fmt.Sprintf("null %s", n.Comment.String())
+	}
+	return n.stringWithoutComment()
+}
+
+func (n *NullNode) stringWithoutComment() string {
 	return "null"
 }
 
@@ -653,6 +679,13 @@ func (n *IntegerNode) GetValue() interface{} {
 
 // String int64 to text
 func (n *IntegerNode) String() string {
+	if n.Comment != nil {
+		return addCommentString(n.Token.Value, n.Comment)
+	}
+	return n.stringWithoutComment()
+}
+
+func (n *IntegerNode) stringWithoutComment() string {
 	return n.Token.Value
 }
 
@@ -694,6 +727,13 @@ func (n *FloatNode) GetValue() interface{} {
 
 // String float64 to text
 func (n *FloatNode) String() string {
+	if n.Comment != nil {
+		return addCommentString(n.Token.Value, n.Comment)
+	}
+	return n.stringWithoutComment()
+}
+
+func (n *FloatNode) stringWithoutComment() string {
 	return n.Token.Value
 }
 
@@ -736,9 +776,48 @@ func (n *StringNode) GetValue() interface{} {
 func (n *StringNode) String() string {
 	switch n.Token.Type {
 	case token.SingleQuoteType:
-		return fmt.Sprintf(`'%s'`, n.Value)
+		quoted := fmt.Sprintf(`'%s'`, n.Value)
+		if n.Comment != nil {
+			return addCommentString(quoted, n.Comment)
+		}
+		return quoted
 	case token.DoubleQuoteType:
-		return strconv.Quote(n.Value)
+		quoted := strconv.Quote(n.Value)
+		if n.Comment != nil {
+			return addCommentString(quoted, n.Comment)
+		}
+		return quoted
+	}
+
+	lbc := token.DetectLineBreakCharacter(n.Value)
+	if strings.Contains(n.Value, lbc) {
+		// This block assumes that the line breaks in this inside scalar content and the Outside scalar content are the same.
+		// It works mostly, but inconsistencies occur if line break characters are mixed.
+		header := token.LiteralBlockHeader(n.Value)
+		space := strings.Repeat(" ", n.Token.Position.Column-1)
+		values := []string{}
+		for _, v := range strings.Split(n.Value, lbc) {
+			values = append(values, fmt.Sprintf("%s  %s", space, v))
+		}
+		block := strings.TrimSuffix(strings.TrimSuffix(strings.Join(values, lbc), fmt.Sprintf("%s  %s", lbc, space)), fmt.Sprintf("  %s", space))
+		return fmt.Sprintf("%s%s%s", header, lbc, block)
+	} else if len(n.Value) > 0 && (n.Value[0] == '{' || n.Value[0] == '[') {
+		return fmt.Sprintf(`'%s'`, n.Value)
+	}
+	if n.Comment != nil {
+		return addCommentString(n.Value, n.Comment)
+	}
+	return n.Value
+}
+
+func (n *StringNode) stringWithoutComment() string {
+	switch n.Token.Type {
+	case token.SingleQuoteType:
+		quoted := fmt.Sprintf(`'%s'`, n.Value)
+		return quoted
+	case token.DoubleQuoteType:
+		quoted := strconv.Quote(n.Value)
+		return quoted
 	}
 
 	lbc := token.DetectLineBreakCharacter(n.Value)
@@ -800,7 +879,15 @@ func (n *LiteralNode) GetValue() interface{} {
 // String literal to text
 func (n *LiteralNode) String() string {
 	origin := n.Value.GetToken().Origin
-	return fmt.Sprintf("%s\n%s", n.Start.Value, strings.TrimRight(strings.TrimRight(origin, " "), "\n"))
+	lit := strings.TrimRight(strings.TrimRight(origin, " "), "\n")
+	if n.Comment != nil {
+		return fmt.Sprintf("%s %s\n%s", n.Start.Value, n.Comment.String(), lit)
+	}
+	return fmt.Sprintf("%s\n%s", n.Start.Value, lit)
+}
+
+func (n *LiteralNode) stringWithoutComment() string {
+	return n.String()
 }
 
 // MarshalYAML encodes to a YAML text
@@ -834,6 +921,10 @@ func (n *MergeKeyNode) GetValue() interface{} {
 
 // String returns '<<' value
 func (n *MergeKeyNode) String() string {
+	return n.Token.Value
+}
+
+func (n *MergeKeyNode) stringWithoutComment() string {
 	return n.Token.Value
 }
 
@@ -879,6 +970,13 @@ func (n *BoolNode) GetValue() interface{} {
 
 // String boolean to text
 func (n *BoolNode) String() string {
+	if n.Comment != nil {
+		return addCommentString(n.Token.Value, n.Comment)
+	}
+	return n.stringWithoutComment()
+}
+
+func (n *BoolNode) stringWithoutComment() string {
 	return n.Token.Value
 }
 
@@ -919,6 +1017,13 @@ func (n *InfinityNode) GetValue() interface{} {
 
 // String infinity to text
 func (n *InfinityNode) String() string {
+	if n.Comment != nil {
+		return addCommentString(n.Token.Value, n.Comment)
+	}
+	return n.stringWithoutComment()
+}
+
+func (n *InfinityNode) stringWithoutComment() string {
 	return n.Token.Value
 }
 
@@ -958,6 +1063,13 @@ func (n *NanNode) GetValue() interface{} {
 
 // String returns .nan
 func (n *NanNode) String() string {
+	if n.Comment != nil {
+		return addCommentString(n.Token.Value, n.Comment)
+	}
+	return n.stringWithoutComment()
+}
+
+func (n *NanNode) stringWithoutComment() string {
 	return n.Token.Value
 }
 
@@ -1064,34 +1176,61 @@ func (n *MappingNode) AddColumn(col int) {
 	}
 }
 
-func (n *MappingNode) flowStyleString() string {
-	if len(n.Values) == 0 {
-		return "{}"
-	}
+func (n *MappingNode) flowStyleString(commentMode bool) string {
 	values := []string{}
 	for _, value := range n.Values {
 		values = append(values, strings.TrimLeft(value.String(), " "))
 	}
-	return fmt.Sprintf("{%s}", strings.Join(values, ", "))
+	mapText := fmt.Sprintf("{%s}", strings.Join(values, ", "))
+	if commentMode && n.Comment != nil {
+		return addCommentString(mapText, n.Comment)
+	}
+	return mapText
 }
 
-func (n *MappingNode) blockStyleString() string {
-	if len(n.Values) == 0 {
-		return "{}"
-	}
+func (n *MappingNode) blockStyleString(commentMode bool) string {
 	values := []string{}
 	for _, value := range n.Values {
 		values = append(values, value.String())
 	}
-	return strings.Join(values, "\n")
+	mapText := strings.Join(values, "\n")
+	if commentMode && n.Comment != nil {
+		value := values[0]
+		var spaceNum int
+		for i := 0; i < len(value); i++ {
+			if value[i] != ' ' {
+				break
+			}
+			spaceNum++
+		}
+		comment := n.Comment.StringWithSpace(spaceNum)
+		return fmt.Sprintf("%s\n%s", comment, mapText)
+	}
+	return mapText
 }
 
 // String mapping values to text
 func (n *MappingNode) String() string {
-	if n.IsFlowStyle || len(n.Values) == 0 {
-		return n.flowStyleString()
+	if len(n.Values) == 0 {
+		if n.Comment != nil {
+			return addCommentString("{}", n.Comment)
+		}
+		return "{}"
 	}
-	return n.blockStyleString()
+
+	commentMode := true
+	if n.IsFlowStyle || len(n.Values) == 0 {
+		return n.flowStyleString(commentMode)
+	}
+	return n.blockStyleString(commentMode)
+}
+
+func (n *MappingNode) stringWithoutComment() string {
+	commentMode := false
+	if n.IsFlowStyle || len(n.Values) == 0 {
+		return n.flowStyleString(commentMode)
+	}
+	return n.blockStyleString(commentMode)
 }
 
 // MapRange implements MapNode protocol
@@ -1137,6 +1276,10 @@ func (n *MappingKeyNode) AddColumn(col int) {
 
 // String tag to text
 func (n *MappingKeyNode) String() string {
+	return fmt.Sprintf("%s %s", n.Start.Value, n.Value.String())
+}
+
+func (n *MappingKeyNode) stringWithoutComment() string {
 	return fmt.Sprintf("%s %s", n.Start.Value, n.Value.String())
 }
 
@@ -1199,6 +1342,64 @@ func (n *MappingValueNode) SetIsFlowStyle(isFlow bool) {
 
 // String mapping value to text
 func (n *MappingValueNode) String() string {
+	if n.Comment != nil {
+		return fmt.Sprintf(
+			"%s\n%s",
+			n.Comment.StringWithSpace(n.Key.GetToken().Position.Column-1),
+			n.toString(),
+		)
+	}
+	return n.toString()
+}
+
+func (n *MappingValueNode) toString() string {
+	space := strings.Repeat(" ", n.Key.GetToken().Position.Column-1)
+	keyIndentLevel := n.Key.GetToken().Position.IndentLevel
+	valueIndentLevel := n.Value.GetToken().Position.IndentLevel
+	keyComment := n.Key.GetComment()
+	if _, ok := n.Value.(ScalarNode); ok {
+		return fmt.Sprintf("%s%s: %s", space, n.Key.String(), n.Value.String())
+	} else if keyIndentLevel < valueIndentLevel {
+		if keyComment != nil {
+			return fmt.Sprintf(
+				"%s%s: %s\n%s",
+				space,
+				n.Key.stringWithoutComment(),
+				keyComment.String(),
+				n.Value.String(),
+			)
+		}
+		return fmt.Sprintf("%s%s:\n%s", space, n.Key.String(), n.Value.String())
+	} else if m, ok := n.Value.(*MappingNode); ok && (m.IsFlowStyle || len(m.Values) == 0) {
+		return fmt.Sprintf("%s%s: %s", space, n.Key.String(), n.Value.String())
+	} else if s, ok := n.Value.(*SequenceNode); ok && (s.IsFlowStyle || len(s.Values) == 0) {
+		return fmt.Sprintf("%s%s: %s", space, n.Key.String(), n.Value.String())
+	} else if _, ok := n.Value.(*AnchorNode); ok {
+		return fmt.Sprintf("%s%s: %s", space, n.Key.String(), n.Value.String())
+	} else if _, ok := n.Value.(*AliasNode); ok {
+		return fmt.Sprintf("%s%s: %s", space, n.Key.String(), n.Value.String())
+	}
+	if keyComment != nil {
+		return fmt.Sprintf(
+			"%s%s: %s\n%s",
+			space,
+			n.Key.stringWithoutComment(),
+			keyComment.String(),
+			n.Value.String(),
+		)
+	}
+	if m, ok := n.Value.(*MappingNode); ok && m.Comment != nil {
+		return fmt.Sprintf(
+			"%s%s: %s",
+			space,
+			n.Key.String(),
+			strings.TrimLeft(n.Value.String(), " "),
+		)
+	}
+	return fmt.Sprintf("%s%s:\n%s", space, n.Key.String(), n.Value.String())
+}
+
+func (n *MappingValueNode) stringWithoutComment() string {
 	space := strings.Repeat(" ", n.Key.GetToken().Position.Column-1)
 	keyIndentLevel := n.Key.GetToken().Position.IndentLevel
 	valueIndentLevel := n.Value.GetToken().Position.IndentLevel
@@ -1263,10 +1464,11 @@ func (m *ArrayNodeIter) Len() int {
 // SequenceNode type of sequence node
 type SequenceNode struct {
 	*BaseNode
-	Start       *token.Token
-	End         *token.Token
-	IsFlowStyle bool
-	Values      []Node
+	Start         *token.Token
+	End           *token.Token
+	IsFlowStyle   bool
+	Values        []Node
+	ValueComments []*CommentGroupNode
 }
 
 // Replace replace value node.
@@ -1340,7 +1542,11 @@ func (n *SequenceNode) flowStyleString() string {
 func (n *SequenceNode) blockStyleString() string {
 	space := strings.Repeat(" ", n.Start.Position.Column-1)
 	values := []string{}
-	for _, value := range n.Values {
+	if n.Comment != nil {
+		values = append(values, n.Comment.StringWithSpace(n.Start.Position.Column-1))
+	}
+
+	for idx, value := range n.Values {
 		valueStr := value.String()
 		splittedValues := strings.Split(valueStr, "\n")
 		trimmedFirstValue := strings.TrimLeft(splittedValues[0], " ")
@@ -1356,6 +1562,9 @@ func (n *SequenceNode) blockStyleString() string {
 			newValues = append(newValues, fmt.Sprintf("%s  %s", space, trimmed))
 		}
 		newValue := strings.Join(newValues, "\n")
+		if len(n.ValueComments) == len(n.Values) && n.ValueComments[idx] != nil {
+			values = append(values, n.ValueComments[idx].StringWithSpace(n.Start.Position.Column-1))
+		}
 		values = append(values, fmt.Sprintf("%s- %s", space, newValue))
 	}
 	return strings.Join(values, "\n")
@@ -1363,6 +1572,13 @@ func (n *SequenceNode) blockStyleString() string {
 
 // String sequence to text
 func (n *SequenceNode) String() string {
+	if n.IsFlowStyle || len(n.Values) == 0 {
+		return n.flowStyleString()
+	}
+	return n.blockStyleString()
+}
+
+func (n *SequenceNode) stringWithoutComment() string {
 	if n.IsFlowStyle || len(n.Values) == 0 {
 		return n.flowStyleString()
 	}
@@ -1439,6 +1655,10 @@ func (n *AnchorNode) String() string {
 	return fmt.Sprintf("&%s %s", n.Name.String(), value)
 }
 
+func (n *AnchorNode) stringWithoutComment() string {
+	return n.String()
+}
+
 // MarshalYAML encodes to a YAML text
 func (n *AnchorNode) MarshalYAML() ([]byte, error) {
 	return []byte(n.String()), nil
@@ -1489,6 +1709,10 @@ func (n *AliasNode) String() string {
 	return fmt.Sprintf("*%s", n.Value.String())
 }
 
+func (n *AliasNode) stringWithoutComment() string {
+	return fmt.Sprintf("*%s", n.Value.String())
+}
+
 // MarshalYAML encodes to a YAML text
 func (n *AliasNode) MarshalYAML() ([]byte, error) {
 	return []byte(n.String()), nil
@@ -1523,6 +1747,10 @@ func (n *DirectiveNode) AddColumn(col int) {
 
 // String directive to text
 func (n *DirectiveNode) String() string {
+	return fmt.Sprintf("%s%s", n.Start.Value, n.Value.String())
+}
+
+func (n *DirectiveNode) stringWithoutComment() string {
 	return fmt.Sprintf("%s%s", n.Start.Value, n.Value.String())
 }
 
@@ -1564,6 +1792,10 @@ func (n *TagNode) String() string {
 	return fmt.Sprintf("%s %s", n.Start.Value, n.Value.String())
 }
 
+func (n *TagNode) stringWithoutComment() string {
+	return fmt.Sprintf("%s %s", n.Start.Value, n.Value.String())
+}
+
 // MarshalYAML encodes to a YAML text
 func (n *TagNode) MarshalYAML() ([]byte, error) {
 	return []byte(n.String()), nil
@@ -1572,6 +1804,7 @@ func (n *TagNode) MarshalYAML() ([]byte, error) {
 // CommentNode type of comment node
 type CommentNode struct {
 	*BaseNode
+	Token *token.Token
 }
 
 // Read implements (io.Reader).Read
@@ -1583,20 +1816,84 @@ func (n *CommentNode) Read(p []byte) (int, error) {
 func (n *CommentNode) Type() NodeType { return CommentType }
 
 // GetToken returns token instance
-func (n *CommentNode) GetToken() *token.Token { return n.Comment }
+func (n *CommentNode) GetToken() *token.Token { return n.Token }
 
 // AddColumn add column number to child nodes recursively
 func (n *CommentNode) AddColumn(col int) {
-	n.Comment.AddColumn(col)
+	if n.Token == nil {
+		return
+	}
+	n.Token.AddColumn(col)
 }
 
 // String comment to text
 func (n *CommentNode) String() string {
-	return n.Comment.Value
+	return fmt.Sprintf("#%s", n.Token.Value)
+}
+
+func (n *CommentNode) stringWithoutComment() string {
+	return ""
 }
 
 // MarshalYAML encodes to a YAML text
 func (n *CommentNode) MarshalYAML() ([]byte, error) {
+	return []byte(n.String()), nil
+}
+
+// CommentGroupNode type of comment node
+type CommentGroupNode struct {
+	*BaseNode
+	Comments []*CommentNode
+}
+
+// Read implements (io.Reader).Read
+func (n *CommentGroupNode) Read(p []byte) (int, error) {
+	return readNode(p, n)
+}
+
+// Type returns TagType
+func (n *CommentGroupNode) Type() NodeType { return CommentType }
+
+// GetToken returns token instance
+func (n *CommentGroupNode) GetToken() *token.Token {
+	if len(n.Comments) > 0 {
+		return n.Comments[0].Token
+	}
+	return nil
+}
+
+// AddColumn add column number to child nodes recursively
+func (n *CommentGroupNode) AddColumn(col int) {
+	for _, comment := range n.Comments {
+		comment.AddColumn(col)
+	}
+}
+
+// String comment to text
+func (n *CommentGroupNode) String() string {
+	values := []string{}
+	for _, comment := range n.Comments {
+		values = append(values, comment.String())
+	}
+	return strings.Join(values, "\n")
+}
+
+func (n *CommentGroupNode) StringWithSpace(col int) string {
+	space := strings.Repeat(" ", col)
+	values := []string{}
+	for _, comment := range n.Comments {
+		values = append(values, space+comment.String())
+	}
+	return strings.Join(values, "\n")
+
+}
+
+func (n *CommentGroupNode) stringWithoutComment() string {
+	return ""
+}
+
+// MarshalYAML encodes to a YAML text
+func (n *CommentGroupNode) MarshalYAML() ([]byte, error) {
 	return []byte(n.String()), nil
 }
 
@@ -1665,6 +1962,77 @@ func (v *filterWalker) Visit(n Node) Visitor {
 		v.results = append(v.results, n)
 	}
 	return v
+}
+
+type parentFinder struct {
+	target Node
+}
+
+func (f *parentFinder) walk(parent, node Node) Node {
+	if f.target == node {
+		return parent
+	}
+	switch n := node.(type) {
+	case *CommentNode:
+		return nil
+	case *NullNode:
+		return nil
+	case *IntegerNode:
+		return nil
+	case *FloatNode:
+		return nil
+	case *StringNode:
+		return nil
+	case *MergeKeyNode:
+		return nil
+	case *BoolNode:
+		return nil
+	case *InfinityNode:
+		return nil
+	case *NanNode:
+		return nil
+	case *LiteralNode:
+		return f.walk(node, n.Value)
+	case *DirectiveNode:
+		return f.walk(node, n.Value)
+	case *TagNode:
+		return f.walk(node, n.Value)
+	case *DocumentNode:
+		return f.walk(node, n.Body)
+	case *MappingNode:
+		for _, value := range n.Values {
+			if found := f.walk(node, value); found != nil {
+				return found
+			}
+		}
+	case *MappingKeyNode:
+		return f.walk(node, n.Value)
+	case *MappingValueNode:
+		if found := f.walk(node, n.Key); found != nil {
+			return found
+		}
+		return f.walk(node, n.Value)
+	case *SequenceNode:
+		for _, value := range n.Values {
+			if found := f.walk(node, value); found != nil {
+				return found
+			}
+		}
+	case *AnchorNode:
+		if found := f.walk(node, n.Name); found != nil {
+			return found
+		}
+		return f.walk(node, n.Value)
+	case *AliasNode:
+		return f.walk(node, n.Value)
+	}
+	return nil
+}
+
+// Parent get parent node from child node.
+func Parent(root, child Node) Node {
+	finder := &parentFinder{target: child}
+	return finder.walk(root, root)
 }
 
 // Filter returns a list of nodes that match the given type.